Add fuzzy option to show command.
When "show" command is invoked with --fuzzy or -f, the last arguments
are treated as search words instead of the full password path.
Passwords whose paths contain all given words (as substrings) are
listed.  If only one match is found, the orignal "show" command is
executed with the found path.